LayoutTests/fullscreen/full-screen-request.html

<body>
<script src="full-screen-test.js"></script>
<script>
    // Bail out early if the full screen API is not enabled or is missing:
    if (Element.prototype.webkitRequestFullScreen == undefined) {
        logResult(false, "Element.prototype.webkitRequestFullScreen == undefined");
        endTest();
    } else {
        waitForEventTestAndEnd(document, 'webkitfullscreenchange', "document.webkitFullScreen==true");
        document.documentElement.webkitRequestFullScreen();
    }
</script>
