LayoutTests/fullscreen/full-screen-request-expected.txt

EVENT(webkitfullscreenchange) TEST(document.webkitFullScreen==true) OK
END OF TEST
